[sui-node] post-fork startup and metrics (#22862)
## Description 

Today when a fork happens, the sui-node process just crashloops,
attempting to start up but forking and panicing repeatedly until the
systemd unit or kube controller stops restarting it. This makes it hard
for us to know if a node in the network that someone else is running has
panicked. all we know is they have stopped pushing us metrics.

This change includes the following:
* When a fork first happens we commit a fork marker to the
CheckpointWatermark table, then panic
* Adds a check for the fork marker to node startup, supporting a "crash
mode" which emits metrics and logs the error that caused the crash, but
does not continue crashing

## Test plan 

basic unit tests

local cluster tests TODO

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: