Fork recovery: clear local checkpoints on validator start (#23343)
## Description 

Instead of applying checkpoint overrides in CheckpointBuilder, we simply
detect and clear invalid local checkpoints when the node starts. This
handles both the case when we rebuild a local checkpoint with an
different result and the case where we insert a verified checkpoint with
a different digest than a locally built checkpoint.

## Test plan 

Simtests pass, new UT. verified with one-off devnet test for both txn
fork detected and checkpoint fork detected.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: