tokio runtime fix (#23370)
## Description 
The key change in in data_store.rs and it's picking a proper tokio
runtime. The code before was hanging in mysterious ways after a number
of transactions. After investigations the hang appeared related to tokio
and the runtime picked to block on. The code as is now executes
thousands of transactions and it seems stable.
There is a bunch of logging changes that I am adding because they helped
me diagnose the issue with the hang/deadlock. Also they provide a good
overview of what is going on. So I am leaving them here for now. The
file `summary_metrics.rs` si questionable but it is ok for now in my
opinion

## Test plan 

Run a subset of a few thousands transactions to verify code works

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: