[2/n][sui-adapter] Update (but dont enable) new MoveLoc gas charging (#23232)
## Description 

What it says on the tin -- `MoveLoc` doesn't actually move the whole
value (as a slab of bytes) to the stack, but instead the reference to
it, so only charge for the reference size.

This PR only adds the change but it is gated and not enabled in this PR.
The enabling of this will be done at a later date/alongside other
changes.

## Test plan 

CI -- nothing should change. 

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: