fix(errors): Don't use debug format for errors
## Description

`SuiErrors` are user-facing. As much as possible, we should avoid using `Debug` formats for the constituents of these errors, because they may make it to developers.

This also fixes a CI breakage in `main` due to an error being wrapped in another error that displayed it using its Debug format.

## Test plan
CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to your changes, release notes aren't required.

For each box you select, include information after the relevant heading that describes the impact of your changes that a user might notice and any actions they must take to implement updates. 

- [ ] Protocol: 
- [x] Nodes (Validators and Full nodes): Errors will use human-readable descriptions where possible.
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: