Add reconfigure to balance withdraw scheduler (#23536)
## Description 

This PR adds reconfigure API to balance withdraw scheduler.
This is necessary because if a node promotes from fullnode to validator
during epoch change, we need to set the last settled version at that
time, since when the node is running as fullnode we never settle.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: