graphql-alt: add x-sui-rpc-request-id response header (#23604)
## Description 

Adds a `x-sui-rpc-request-id` response header that returns the
`Session.request_id` for debugging.

## Test plan 

Adds new tests that assert a `request_id` is returned (value is not
asserted since it is random) for both successful and failed requests.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: