[sui-indexer-alt-framework] Address `first_checkpoint` + pipeline with no watermark row commit stall (#23209)
## Description

Currently, if `first_checkpoint` is configured and a pipeline with no
watermark row is added to the indexer, it'll essentially stall since the
pipeline's committer task creates a default watermark with
checkpoint_hi_inclusive = 0 and waits for this checkpoint. This prevents
someone from using `first_checkpoint` to arbitrarily seed the starting
point of pipelines for the first time. Consequently, indexer operators
have to manually update the watermarks table to set a starting point
that isn't 0.

In this PR, the indexer is now responsible for determining the
`next_checkpoint` to commit and make watermark updates for each
pipeline. In practice, for a valid `first_checkpoint`
(`next_checkpoint`), sequential pipelines will wait to commit the next
processed checkpoint after its watermark, while concurrent pipelines
will start from the same `first_checkpoint`.

Meanwhile, the pipeline tasks initialize a default watermark variable
for the logger, but in practice will always be overwritten by the next
processed checkpoint for metrics and other reporting.

## Test plan 

- Unit tests for the behavior above
- commit_watermark and sequential committer tests 

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: