sui-core: track and store unchanged loaded runtime objects (#23716)
This patch implements tracking of unchanged loaded runtime objects and
stores this information in the perpetual store. This is done so that we
can package up a checkpoint that includes full information necessary to
either perform any analysis based on the objects read or written by a
transaction or to efficiently replay a transaction without needing to
fetch any additional objects (except packages).

The tracking is done by wrapping the BackingStore that is provided to
the VM with one that tracks and saves all objects that were loaded
during execution. This is more accurate than the existing
"loaded_runtime_objects" structure returned from the execution layer
since it also will capture things like the coin deny list.

In addition to this a new `Checkpoint` structure is introduced that is
intended to eventually replace the existing `CheckpointData` structure.
This new type is intended to only be used in-memory (if you need a
persisted structure the proto version of the message should instead be
used) and it explicitly does not implement Serialize/Deserialize so that
it cannot be accidentally be persisted.