annotated-visitor: Expose layout everywhere
## Description

Add a function to all drivers to return the layout for the Moev value
currently being visited. This will be used by the extractor visitor in
Display v2 which needs to produce a pair of some bytes and an associated
layout when it is slicing out a sub-value after performing some nested
accesses.

## Test plan

CI