docs: update on the blackjack documentation to match refactor changes (#23885)
## Description 

Updates on the blackjack documentation to match new implementation of
the mini game (counter nft was removed and on-chain randomness added on
its place).

## Test plan 

N/A
---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: