Introduce display registry object `0xd` & snapshots (#23922)
## Description 

Introduce all the logic (but not enable) for the creation of Display
Registry (`0xd`).

This will land before #23710. On that PR, the feature flag will be
enabled & will have e2e tests.

## Test plan 

Just testing that the existing tests pass. 

**This PR does not do the creation of the object, only sets up the types
& logic so we can release in two steps (to avoid breaking indexing)**.
Tests for that will be introduced in the PR that enables the feature
flag, because the move code is not yet landed (so creation cannot really
happen yet!)

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: