Output and metrics in the replay (#23933)
## Description 

Small change in the output of the replay, adding println for every
transaction executed
\> Replayed txn 3a5rwWp4dnKQPsxmhuRCb8cn9zp7MsoeswVM8BMz9Rpa (OK):
exec_ms=47 total_ms=670
that is with no flags and no logs
Also started to move code around, in this pr metrics for execution are
the most change. More coming

## Test plan 

I tried it

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: