[move] Capture backtrace at point of error creation in debug builds (#24333)
## Description 

Capture a Rust backtrace at the point an error is created in the VM if
the `MOVE_VM_ERROR_LOCATION` environment flag is set (along with
`RUST_BACKTRACE=1`) in _debug builds only_.

See the comments in the code, but we don't want to capture the backtrace
all the time in debug builds (even if `RUST_BACKTRACE` is set) as this
would e.g., break CI jobs that run with `RUST_BACKTRACE` set since we
use the debug format for VM errors quite a bit.


## Test plan 

CI + tested it out locally.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] Indexing Framework: